requirements.txt
feedparser
psycopg2-binary

